requirements.txt
2to3==1.0
astroid==2.0.4
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
Django==2.1.1
djangorestframework==3.8.2
feedparser==5.2.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.13
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.5
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
